backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
langchain
langgraph
langchain-core
langchain-community
langchain-google-genai
pydantic
